# docs/registers/openapi.yaml
...
openapi: 3.0.3
info:
  title: RFS Station Manager API
  description: |
    REST API for the RFS Station Manager - a real-time digital sign-in system for Rural Fire Service stations.
    
    This API provides endpoints for:
    - Member management
    - Activity tracking
    - Check-in/check-out operations
    - Event management
    - Truck checks
    - Achievements
    
    Real-time updates are also available via Socket.io WebSocket connections.
  version: 1.0.0
  contact:
    name: RFS Station Manager Support
    url: https://github.com/richardthorek/Station-Manager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://bungrfsstation.azurewebsites.net
    description: Production server
...
paths:
...
  /api/organizations/current/invites:
    post:
      tags:
        - Organizations
      summary: Create a multi-use org invite link
      description: Full schema in docs/registers/api_register.json → endpoints.organizations.
      operationId: createOrgInvite
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Invite created
    get:
      tags:
        - Organizations
      summary: List org invite links
      operationId: listOrgInvites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invite list
  /api/organizations/current/members:
    get:
      tags:
        - Organizations
      summary: List org members (multi-org memberships)
      operationId: listOrgMembers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Member list
...
components:
...
  securitySchemes:
    # Future: JWT Bearer token authentication for admin functions
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# No security required currently (by design for kiosk access)
# Future versions may require authentication for admin operations
security: []